TST: Reorganizing test_function.py module into unit and integration tests.
